\docType{data}
\name{Adducts}
\alias{Adducts}
\title{Default adduct lists used by \code{findMAIN}}
\format{A list of two character vectors:
\itemize{
  \item Positive. default adducts used in ESI(+) mode
  \item Negative. default adducts used in ESI(-) mode
}}
\usage{
Adducts
}
\description{
Default adduct lists used by \code{\link{findMAIN}}, serving also as a
template for custom lists.
}
\keyword{datasets}
\keyword{internal}