\name{chemical_elements}
\alias{chemical_elements}
\docType{data}
\title{
chemical_elements
}
\description{
List of chemical elements.
}
\usage{data("chemical_elements")}
\format{
  A data frame with 103 observations on the following 2 variables.
  \describe{
    \item{\code{name}}{a character vector}
    \item{\code{mass}}{a numeric vector}
  }
}
\examples{
data(chemical_elements)
head(chemical_elements)
}
\keyword{datasets}
\keyword{internal}