% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountChemicalElements.R
\name{CountChemicalElements}
\alias{CountChemicalElements}
\title{CountChemicalElements.}
\usage{
CountChemicalElements(x = NULL, ele = NULL)
}
\arguments{
\item{x}{Chemical formula.}

\item{ele}{Character vector of elements to count particularly or counting all contained if NULL.}
}
\value{
A named numeric with counts for all contained or specified elements.
}
\description{
\code{CountChemicalElements} will split a character (chemical formula) 
    into its elements and count their occurrence.
}
\details{
No testing for any chemical alphabet is performed. Elements may occur 
    several times and will be summed up in this case without a warning.
}
