% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findiso.R
\name{findiso}
\alias{findiso}
\title{findiso.}
\usage{
findiso(spec = NULL, mzabs = 0.001, intthr = 0.03, CAMERAlike = TRUE)
}
\arguments{
\item{spec}{A two-column matrix with mz and int.}

\item{mzabs}{Expected maximum within scan mass defect of your device in Dalton.}

\item{intthr}{Do not consider isomain peaks below this intensity cutoff (relative to base peak).}

\item{CAMERAlike}{T/F reformatting the result similar to a CAMERA isocluster search output.}
}
\value{
An annotated spectrum.
}
\description{
\code{findiso} will evaluate a mass spectrum and try to find the main isotopic clusters and adducts.
}
\details{
This function is used within \link{findMAIN} to identify main isotopic clusters.
}
\keyword{internal}
