% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_exactmass.R
\name{get_exactmass}
\alias{get_exactmass}
\title{get_exactmass.}
\usage{
get_exactmass(x)
}
\arguments{
\item{x}{Vector of chemical formulas.}
}
\value{
A named vector of exact masses.
}
\description{
Get the exact mass for chemical formulas.
}
\examples{
InterpretMSSpectrum::get_exactmass(c("C6H12O6", "Na", "H1"))

}
