% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetGroupFactor.R
\name{GetGroupFactor}
\alias{GetGroupFactor}
\title{GetGroupFactor.}
\usage{
GetGroupFactor(x, gap)
}
\arguments{
\item{x}{numeric vector.}

\item{gap}{difference up from which a new group is assumed.}
}
\value{
factor of length x, with groups of x which are seperated by a value >gap.
}
\description{
\code{GetGroupFactor} will provide a factor vector which allows to split any numeric vector
into Groups according to a gap parameter for further processing.
}
\details{
not exported
}
\keyword{internal}

