% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetIsotopeDistribution.R
\name{GetIsotopeDistribution}
\alias{GetIsotopeDistribution}
\title{GetIsotopeDistribution.}
\usage{
GetIsotopeDistribution(fml = NULL, res = NULL, n = 2, ele_vec = c("C",
  "H", "N", "O", "P", "S", "Si"))
}
\arguments{
\item{fml}{sum formula.}

\item{res}{MS resolution. Yet experimental, may fail.}

\item{n}{Number of isotopes to calculate.}

\item{ele_vec}{Character vector of elements to consider.}
}
\value{
Isotope distribution formatted similar to Rdisop result but more precise using enviPat.
}
\description{
\code{GetIsotopeDistribution} will generate an isotopic distribution for a given formula.
}
\details{
not exported
}
\keyword{internal}

