\docType{data}
\name{OrbiMS1}
\alias{OrbiMS1}
\title{Orbitrap MS1 spectra}
\format{A list with 550 matrices (spectra). Two attributes are attached to 
  each spectrum: \itemize{ \item Formula. sum formula of (neutral) compound 
  \item ExactMass. exact mass of (neutral) compound }}
\usage{
OrbiMS1
}
\description{
A set of 550 MS1 pseudospectra of metabolite standards, acquired on an 
Orbitrap-type mass analyzer (Q Exactive, Thermo-Fisher) in electrospray
ionization (ESI) positive mode. Spectra were generated from Thermo raw files
using xcms/CAMERA.
}
\keyword{datasets}
\keyword{internal}