% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadSpecClipboard.R
\name{ReadSpecClipboard}
\alias{ReadSpecClipboard}
\title{ReadSpecClipboard.}
\usage{
ReadSpecClipboard()
}
\value{
A spectrum as two-column matrix.
}
\description{
Read a mass spectrum from the windows clipboard.
}
\examples{
\dontrun{
  if (length(grep("Windows", sessionInfo()$running))==1) {
    x <- InterpretMSSpectrum::apci_spectrum
    write.table(x, "clipboard", sep="\t", row.names=FALSE)
    InterpretMSSpectrum::ReadSpecClipboard()
  }
}
}
