% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Adducts}
\alias{Adducts}
\title{Default adduct lists used by \code{findMAIN}.}
\format{
A list of two character vectors:
\describe{
\item{Positive}{default adducts used in ESI(+) mode}
\item{Negative}{default adducts used in ESI(-) mode}
}
}
\source{
A reasonable selection of frequent adducts based on the list in R-package CAMERA
}
\usage{
data(Adducts)
}
\description{
Default adduct lists used by \code{findMAIN}.
}
\keyword{datasets}
