% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{apci_spectrum}
\alias{apci_spectrum}
\title{APCI spectrum}
\format{
A data frame with 47 observations on the following 2 variables: 
\describe{
\item{mz}{a numeric vector ion masses}
\item{int}{a numeric vector of intensities}
}
}
\usage{
data(apci_spectrum)
}
\description{
Example spectrum of Glutamic acid (3TMS) measured on a Bruker impact II.
}
\keyword{datasets}
