% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chemical_elements}
\alias{chemical_elements}
\title{List of chemical elements.}
\format{
A data frame with 103 observations on the following 2 variables:
\describe{
\item{name}{a character vector of elemental abbreviations}
\item{mass}{a numeric vector of exact masses of the elements main isotope}
}
}
\usage{
data(chemical_elements)
}
\description{
List of chemical elements.
}
\keyword{datasets}
