% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{esi_spectrum}
\alias{esi_spectrum}
\title{ESI spectrum}
\format{
A data frame with 42 observations on the following 2 variables: 
\describe{
\item{mz}{a numeric vector ion masses}
\item{int}{a numeric vector of intensities}
}
}
\usage{
data(esi_spectrum)
}
\description{
Example spectrum of Sedoheptulose 7-phosphate measured on a Bruker impact II.
}
\keyword{datasets}
