% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetGroupFactor.R
\name{GetGroupFactor}
\alias{GetGroupFactor}
\title{GetGroupFactor.}
\usage{
GetGroupFactor(x, gap)
}
\arguments{
\item{x}{Numeric vector.}

\item{gap}{Difference between two consecutive values at which a split is generated.}
}
\value{
A factor vector of length(x) indicating the different groups in x.
}
\description{
\code{GetGroupFactor} will split a numeric vector according to a specified gap value. This is often a useful tool and therefore exported to the namespace.
}
\examples{
x <- c(1:3,14:12,6:9)
GetGroupFactor(x=x, gap=2)
split(x, GetGroupFactor(x=x, gap=2))

}
